create CalciteDataContext
